# Reduce micromatch overhead in jest-haste-map HasteFS