@lencioni 2020-06-05
I think this actually suffers from the same bug as https://github.com/facebook/jest/pull/10131