@lencioni 2020-06-08
I'm going to fold this PR into https://github.com/facebook/jest/pull/10131